Add find_partial_seq.py
Mifish provides two types of databases: one with complete+partial genomes and another with only complete genomes. In the former, the sequences have varying lengths, significantly impacting the clustering in UMAP. The latter, while offering complete genomes, provides less information. Therefore, an attempt is made to subtract the latter from the former to identify the partial components. The expectation is to combine this with the 12S fragment obtained from the latter, creating sequences with similar lengths while preserving the entirety of the information.